Implement a compiler for arithmetic exprs (!)
This is all a fairly straight port of the C, meaning it uses a lot of
module-level variables. I would like to sort this out a bit, but the
compiler, parser, and scanner are all fairly intertwined, and I need to
think about the shape of the object graph a bit more.