vm: fix function call stack accounting
I had screwed this up pretty badly in the conversion from C to Go (which
has no pointer arithmetic, of course). It turns out the solution was
simpler than I thought it was going to be, which is just: store the
correct return stack index in the call frame, then actually restore it
to that on return.

Also here are some calls to resetStack on runtime errors, which was
definitely screwing me up in the repl while trying to debug the other!